refactor: move `frame_to_duckdb_registrable` to its own module
makes testing cleaner
