+++ packages/vgplot/widget/mosaic_widget/__init__.py
@@ -3,14 +3,18 @@ from __future__ import annotations
 import logging
 import pathlib
 import time
+from typing import TYPE_CHECKING
 
 import anywidget
 import duckdb
-import narwhals as nw
 import pyarrow as pa
 import traitlets
 
-from narwhals.typing import Frame
+from mosaic_widget.frame_interop import frame_to_duckdb_registrable
+
+if TYPE_CHECKING:
+    from narwhals.typing import Frame
+
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
@@ -35,7 +39,7 @@ class MosaicWidget(anywidget.AnyWidget):
         self,
         spec: dict | None = None,
         con: duckdb.DuckDBPyConnection | None = None,
-        data: dict[str, Frame] | None = None,
+        data: dict[str, "Frame"] | None = None,
         *args,
         **kwargs,
     ):
@@ -61,7 +65,7 @@ class MosaicWidget(anywidget.AnyWidget):
         self.spec = spec
         self.con = con
         for name, df in data.items():
-            self.con.register(name, _frame_to_duckdb_registrable(df))
+            self.con.register(name, frame_to_duckdb_registrable(df))
         self.on_msg(self._handle_custom_msg)
 
     def _handle_custom_msg(self, data: dict, buffers: list):
@@ -99,40 +103,3 @@ class MosaicWidget(anywidget.AnyWidget):
             logger.warning(f"DONE. Slow query {uuid} took {total} ms.\n{sql}")
         else:
             logger.info(f"DONE. Query {uuid} took {total} ms.\n{sql}")
-
-
-def _is_frame_native_to_duckdb(frame: Frame) -> bool:
-    """Check if a frame is natively supported by DuckDB to be registered as a virtual table with zero-copy guarantees."""
-
-    # "<class '{module_path}'>" -> "{module_path}" via slicing
-    module_path = str(type(frame))[8:-2]
-    frame_backend = module_path.split(".")[0]
-    backends_with_native_virtual_table_support = {"pandas", "polars", "pyarrow"}
-    return frame_backend in backends_with_native_virtual_table_support
-
-
-def _frame_to_duckdb_registrable(frame: Frame) -> object:
-    """Converts a native dataframe(-like) object to a DuckDB-registrable object.
-
-    If the passed `frame` is one of the backends supported by DuckDB to be registered as a virtual table with zero-copy guarantees,
-    we return the `frame` itself. Otherwise, we convert it to a Narwhals frame and then to an Arrow table. Based on the backend-specific implementation,
-    this may or may not be a zero-copy operation.
-
-    If the passed `frame` is a lazy frame, it is materialized.
-    """
-    if _is_frame_native_to_duckdb(frame):
-        return frame
-
-    # If frame is not natively registrable to DuckDB, we convert it to an Arrow table via Narwhals.
-    # Based on the backend-specific implementation, this may or may not be zero-copy.
-    nw_frame = nw.from_native(frame)
-    logger.warning(
-        f"Converting {type(frame)} to Arrow table for DuckDB registration. This may not be a zero-copy operation."
-    )
-
-    # Some backends like Ibis, PySpark, etc. have lazy-only Narwhals support, so we must materialize them
-    if isinstance(nw_frame, nw.LazyFrame):
-        logger.warning("Materializing lazy frame")
-        nw_frame = nw_frame.collect()
-
-    return nw_frame.to_arrow()
